requirements.txt
flask==2.3.3
openai-whisper==20230314
numpy==1.24.3
scipy==1.10.1
soundfile==0.12.1
torch==2.0.1
tqdm==4.65.0